+++ src/Ext/Rules/Body.cpp
@@ -205,6 +205,9 @@ void RulesExt::ExtData::LoadBeforeTypeData(RulesClass* pThis, CCINIClass* pINI)
 	this->CombatLightDetailLevel.Read(exINI, GameStrings::AudioVisual, "CombatLightDetailLevel");
 	this->LightFlashAlphaImageDetailLevel.Read(exINI, GameStrings::AudioVisual, "LightFlashAlphaImageDetailLevel");
 
+	this->BountyEnablers_Interceptor.Read(exINI, GameStrings::General, "BountyEnablers.Interceptor");
+	this->BountyDisplay_Interceptor.Read(exINI, GameStrings::AudioVisual, "BountyDisplay.Interceptor");
+
 	// Section AITargetTypes
 	int itemsCount = pINI->GetKeyCount("AITargetTypes");
 	for (int i = 0; i < itemsCount; ++i)
@@ -386,6 +389,8 @@ void RulesExt::ExtData::Serialize(T& Stm)
 		.Process(this->WarheadParticleAlphaImageIsLightFlash)
 		.Process(this->CombatLightDetailLevel)
 		.Process(this->LightFlashAlphaImageDetailLevel)
+		.Process(this->BountyEnablers_Interceptor)
+		.Process(this->BountyDisplay_Interceptor)
 		;
 }