+++ docs/Whats-New.md
@@ -466,6 +466,7 @@ New:
 - `Scorch` / `Flamer` fire animation customization (by Starkku)
 - Warheads parasite removal customization (by Starkku)
 - Allow infantry to use land sequences in water (by Starkku)
+- Experience and Bounty for Interceptor (by Ollerus)
 
 Vanilla fixes:
 - Allow AI to repair structures built from base nodes/trigger action 125/SW delivery in single player missions (by Trsdy)